+++ pkg/FlexBayes/src/fitBayesianMVHLM.cc
@@ -10,15 +10,15 @@
 
 extern "C" {
 
-void  transformResponseAndPredictorsMV( long * number_groups, 
-                                        long * number_data,
+void  transformResponseAndPredictorsMV( int * number_groups, 
+                                        int * number_data,
                                         bool random_effects,
                                         bool fixed_effects,
                                         CMatrix ** response, 
                                         CMatrix *** random_predictors, 
                                         CMatrix *** fixed_predictors,   
-                                        long * unique_error_Cov, 
-                                        long * dim_error_Cov, 
+                                        int * unique_error_Cov, 
+                                        int * dim_error_Cov, 
 		 			double * error_Cov ) throw( rtErr );
 
 /*
@@ -97,75 +97,75 @@ void  transformResponseAndPredictorsMV( long * number_groups,
          otherwise, the matrices tau2 are returned.
 
 */
-void fitBayesianMVHLM( long * number_groups,
-                     long * number_data,
-                     long * dim_response,
-                     long * dim_beta,
-                     long * dim_gamma,
-                     long * dim_alpha,
+void fitBayesianMVHLM( int * number_groups,
+                     int * number_data,
+                     int * dim_response,
+                     int * dim_beta,
+                     int * dim_gamma,
+                     int * dim_alpha,
 
                      double * data_response,
                      double * data_random,
                      double * data_fixed,
                      double * data_level2,
 
-                     long * total_missingR_data,
-                     long * number_response_missing,
-                     long * response_missing,
+                     int * total_missingR_data,
+                     int * number_response_missing,
+                     int * response_missing,
 
-                     long * total_missingRP_data,
-                     long * number_random_predictors_missing,
-                     long * random_predictors_missing,
+                     int * total_missingRP_data,
+                     int * number_random_predictors_missing,
+                     int * random_predictors_missing,
 
-                     long * total_missingFP_data,
-                     long * number_fixed_predictors_missing,
-                     long * fixed_predictors_missing,
+                     int * total_missingFP_data,
+                     int * number_fixed_predictors_missing,
+                     int * fixed_predictors_missing,
 
-                     long * unique_error_Cov,
-                     long * dim_error_Cov,
+                     int * unique_error_Cov,
+                     int * dim_error_Cov,
                      double * error_Cov,
 
                      double * degreesOfFreedom_likelihood,
-                     long * likelihood_type,
+                     int * likelihood_type,
 
                      double * gammamean,
                      double * gammaCov,
                      double * gammaDF,
-                     long * prior_gamma_type,
+                     int * prior_gamma_type,
 
                      double * betamean,
                      double * betaCov,
                      double * betaDF,
-                     long * prior_beta_type,
+                     int * prior_beta_type,
 
                      double * alphamean,
                      double * alphaCov,
                      double * alphaDF,
-                     long * prior_alpha_type,
+                     int * prior_alpha_type,
 
                      double * sigmaDF,
                      double * sigmaScale,
                      double * sigmaPower,
-                     long * common_sigma,
-                     long * prior_sigma_type,
+                     int * common_sigma,
+                     int * prior_sigma_type,
 
                      double * tauDF,
                      double * tauScale,
                      double * tauPower,
-                     long * prior_tau_type,
+                     int * prior_tau_type,
 
-                     long * read_init_point,
+                     int * read_init_point,
                      double * betaInit,
                      double * gammaInit,
                      double * alphaInit,
                      double * sigma2Init,
                      double * tau2Init,
 
-                     long * burnInLength,
-                     long * simulationsToPerform,
-                     long * sampleFrequency,
+                     int * burnInLength,
+                     int * simulationsToPerform,
+                     int * sampleFrequency,
 
-                     long * print_statistics,
+                     int * print_statistics,
                      double * output_simulations,
 		       double * gibbs_drawing_stats ) throw( rtErr )
 {
@@ -914,15 +914,15 @@ void fitBayesianMVHLM( long * number_groups,
 
     
 
-void  transformResponseAndPredictorsMV( long * number_groups, 
-                                        long * number_data,
+void  transformResponseAndPredictorsMV( int * number_groups, 
+                                        int * number_data,
                                         bool random_effects,
                                         bool fixed_effects,
                                         CMatrix ** response, 
                                         CMatrix *** random_predictors, 
                                         CMatrix *** fixed_predictors,   
-                                        long * unique_error_Cov, 
-                                        long * dim_error_Cov, 
+                                        int * unique_error_Cov, 
+                                        int * dim_error_Cov, 
                                         double * error_Cov ) throw( rtErr )
 {
   //construct appropriate error covariance matrix (if needed)
