+++ pkg/FlexBayes/src/getInitialPointsBayesMVHLM.cc
@@ -15,13 +15,13 @@ extern "C" {
 
 
 void getInitialPointsBayesMVHLM( 
-                     long * number_draws,
-                     long * number_groups,
-                     long * number_data,
-                     long * dim_response,
-                     long * dim_beta,
-                     long * dim_gamma,
-                     long * dim_alpha,
+                     int * number_draws,
+                     int * number_groups,
+                     int * number_data,
+                     int * dim_response,
+                     int * dim_beta,
+                     int * dim_gamma,
+                     int * dim_alpha,
 
                      double * data_random,
                      double * data_fixed,
@@ -31,7 +31,7 @@ void getInitialPointsBayesMVHLM(
                      double * betamean,
                      double * betaCov,
                      double * betaDF,
-                     long * prior_beta_type,
+                     int * prior_beta_type,
 
                      double * gammamean,
                      double * gammaCov,
@@ -43,13 +43,13 @@ void getInitialPointsBayesMVHLM(
 
                      double * sigmaDF,
                      double * sigmaScale,
-                     long * dim_sigmaScale,
-                     long * prior_sigma_type,
-                     long * common_sigma,
+                     int * dim_sigmaScale,
+                     int * prior_sigma_type,
+                     int * common_sigma,
 
                      double * tauDF,
                      double * tauScale,
-                     long * prior_tau_type,
+                     int * prior_tau_type,
 
                      double * output_points ) throw( rtErr )
 {
