+++ pkg/FlexBayes/src/getInitialPointsBhlm.cc
@@ -15,12 +15,12 @@ extern "C" {
 
 
 void getInitialPointsBayesHLM( 
-                     long * number_draws,
-                     long * number_groups,
-                     long * number_data,
-                     long * dim_beta,
-                     long * dim_gamma,
-                     long * dim_alpha,
+                     int * number_draws,
+                     int * number_groups,
+                     int * number_data,
+                     int * dim_beta,
+                     int * dim_gamma,
+                     int * dim_alpha,
 
                      double * random_data,
                      double * fixed_data,
@@ -30,7 +30,7 @@ void getInitialPointsBayesHLM(
                      double * betamean,
                      double * betaCov,
                      double * betaDF,
-                     long * prior_beta_type,
+                     int * prior_beta_type,
 
                      double * gammamean,
                      double * gammaCov,
@@ -42,13 +42,13 @@ void getInitialPointsBayesHLM(
 
                      double * sigmaDF,
                      double * sigmaScale,
-                     long * dim_sigmaScale,
-                     long * prior_sigma_type,
-                     long * common_sigma,
+                     int * dim_sigmaScale,
+                     int * prior_sigma_type,
+                     int * common_sigma,
 
                      double * tauDF,
                      double * tauScale,
-                     long * prior_tau_type,
+                     int * prior_tau_type,
 
                      double * output_points ) throw( rtErr )
 {
